Fix sign-compare

+++ ed_perception/src/annotation_gui.cpp
@@ -307,7 +307,7 @@ public:
             }
             else if (key == 84) // down
             {
-                if (i_possible_types + 1 < possible_types.size())
+                if (i_possible_types + 1 < static_cast<int>(possible_types.size()))
                     ++i_possible_types;
             }
             else if (key == 27) // ESC
